Reformatted code and table output
